Implement the removeEmptyItems option to automatically delete physical items with zero quantity. Closes #26

+++ module/actor-sheet.js
@@ -125,9 +125,10 @@ class DCCActorSheet extends ActorSheet {
     }
 
     // Iterate through items, allocating to containers
+    const removeEmptyItems = sheetData.data.config.removeEmptyItems
     for (const i of inventory) {
       // Remove physical items with zero quantity
-      if (i.quantity && i.quantity <= 0) {
+      if (removeEmptyItems && i.data.quantity !== undefined && i.data.quantity <= 0) {
         this.actor.deleteOwnedItem(i._id, {})
         continue
       }
